requirements.txt
streamlit
pandas
numpy
pdfplumber
python-docx
spacy
nltk
fuzzywuzzy
sentence-transformers
openai
langchain
faiss-cpu
chromadb
plotly
seaborn
matplotlib
supabase
python-dotenv
bcrypt
pyjwt
